prov/efa: Move fabric tests to a dedicated file
Move fabric tests out of domain test file.
