+++ prov/efa/Makefile.include
@@ -164,6 +164,7 @@ nodist_prov_efa_test_efa_unit_test_SOURCES = \
 	prov/efa/test/efa_unit_test_common.c \
 	prov/efa/test/efa_unit_test_domain.c \
 	prov/efa/test/efa_unit_test_rdm_domain.c \
+	prov/efa/test/efa_unit_test_fabric.c \
 	prov/efa/test/efa_unit_test_ep.c \
 	prov/efa/test/efa_unit_test_av.c \
 	prov/efa/test/efa_unit_test_cq.c \
